[PORT FROM R2] atomisp: CSS release Synergy BL 29nov2011
BZ: 17272

This is the CSS release from Synergy baseline of 29nov2011.
It fixes the following issues:
	- Standalone acceleration firmware has no support
	  to get its completion interrupt translated
	- ISP parameter for Gamma correction (K1,K2)
	  should be tunable from ISP parameter file
	- ISP BNR/CNR
	- MACC table index order difference between ISP
	  and USER
	- MACC table segment selection problem

Orig-Change-Id: I2eb9ad577e6949ceb1eebbf0666d8450b4375a1c
Reviewed-on: http://android.intel.com:8080/25729
Tested-by: Ilvonen, Ville <ville.ilvonen@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/28039
Tested-by: Koski, Anttu <anttu.koski@intel.com>

+++ drivers/media/video/atomisp/css/sh_css_types.h
@@ -620,8 +620,8 @@ struct sh_css_de_config {
 };
 
 struct sh_css_gc_config {
-	int gain_k1;
-	int gain_k2;
+	unsigned short gain_k1;
+	unsigned short gain_k2;
 };
 
 struct sh_css_anr_config {
